requirements.txt
customtkinter>=5.2.0
python-dotenv>=1.0.0
pyttsx3>=2.90
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
pvporcupine>=3.0.0
pvrecorder>=1.2.0
google-genai>=1.0.0
openai>=1.12.0